hm1/requirements.txt
xmltodict
dicttoxml
PyYAML
avro
msgpack
protobuf
